+++ packages/@wdio_electron-utils/src/binary/select.ts
